add user checking

+++ model/course.py
@@ -198,6 +198,8 @@ def update_students(user, course, users):
         warning = [*({*[u.obj for u in u_users]} - {*course.students})]
         course.update(pull_all__students=users)
         for user in u_users:
+            if user.obj in warning:
+                continue
             for problem in course.problems:
                 if user == problem.author:
                     problem.delete()
